+++ src/SeqKerneling.java
@@ -4,7 +4,7 @@ public class SeqKerneling {
 
 
 
-    public void test(String fileName, String directory, int[][] kernel){
+    public void test(String fileName, String directory, float[][] kernel){
